fix: check correct hex encoding for `BlockIdentifier` & `BlockIdentifierOrHash` (#374)
**Motivation**
Checking that the hex values for the block field of rpc requests have a
"0x" prefix and returning the correct rpc error if they don't

<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Replace `Deserialize` iml of `BlockIdentifier` &
`BlockIdentifierOrHash` with a `parse` method that returns the correct
`RpcError` upon failure and checks for correct prefixes
* Move `BlockIdentifier` & `BlockIdentifierOrHash` to a separate module
under rpc/types
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #371
