+++ crates/rpc/types/mod.rs
@@ -1,4 +1,5 @@
 pub mod block;
+pub mod block_identifier;
 pub mod fork_choice;
 pub mod payload;
 pub mod receipt;
